Add interval sliders for fine tuning of selection
Wiping the histogram figure and recreating its axis each time was a
quick and dirty solution which worked well enough for single selections,
but this is much too slow for dynamic interactions such as incremental
manipulation, so this was fixed by taking the proper approach and using
observables along with setting the legend's label separately.

Things could be simplified by placing the fft and histogram images on
the same figure as well as initializing it with the sliders and a fully
extended rectangle, but at the moment recreating the fft figure on
histogram window close happens fast enough and it's nice to be able to
independently move and resize the figures as well as clear the rectangle
by closing the histogram window.