Add live update switch
This turns off the live updating of the histogram by default as it can
be slow when analyzing high resolution images on old hardware.
The figure can now be manually updated and the live updating can also
be toggled. Currently this isn't ideal since the figure is presented
using a different window and will defocus when interacting with the fft,
but both images will be set on the same figure window in a subsequent
version.

This design was chosen as mouse events when interacting with the sliders
are not reliable; ideally we'd update the figure when the mouse button
is released, but at the moment nothing is registered when the slider
elements are interacted with by dragging. This is also part of the
reason why the symmetry logic for the sliders is so hacky and slow;
if we could set the observables dynamically and only set the slider
positions once the mouse button is released then the performance would
improve.