Improvements and fixes for the locked sliders
Some slight simplifications and the addition of missing
observer switches when setting the orthogonal sliders under
aspect ratio lock.

This implementation is far from ideal as it is still relatively slow,
requires setting the sliders again directly on every slider interaction,
requires flipping the observer functions to prevent recursive behavior
and overflow, and the selection will sometimes decenter.
Setting the sliders so frequently also results in more calls to the
refresh function which makes the histogram update slower.

The reason why interval sliders were chosen over the normal sliders was
for more fine grained control. It's a nice feature to be able to move
the rectangle around by dragging the interval slider, and with live
updating this allows us to see how a section of the fft changes
dynamically. However, since we're really only interested in symmetric
changes to the rectangle it makes little sense to change the position of
only one corner point and there are too many sacrifices being made to
justify interval sliders over regular sliders. Section analysis can be
done through selection anyway, just not dynamically, so the sliders will
most likely be changed in a subsequent version.