Simplify everything
This rewrite rearranges things so that both images (the fft and the
histogram) are set on the same figure and all elements are initialized
at the beginning. Previous versions were convoluted because essentially
we had to keep track of 3 states: the initial state, the state where a
selection is made, and the state with the histogram window closed.

This makes the script much easier to maintain and use.