Clean up the DataInspector tooltips
Although the latest version of Makie provides an updated API for
modifying and creating custom tooltips using the inspector_label
attribute on a per plot basis, in practice this produces an error
related to gl_convert. As such, an easy workaround is to simply override
the methods used to construct the strings for display. This is probably
a better approach anyway since the histogram is created and destroyed
each time the Live toggle is toggled so this saves us from having to set
the attribute each time.