Add save buttons for the figure and FFT
Once again kind of hacky, but it gets the job done. This makes things a
bit more user friendly and allows us to avoid the type piracy on show
because we no longer need to return the FFT matrix.

colorbuffer here returns the screen as an RGB matrix and allows us to
save the figures while preserving the main window. For the FFT a new
invisible screen is created and then destroyed after saving. Using the
Screen constructor directly (without `display`) requires a Scene type
and passing it the Figure's scene member seems to route in a way which
ends up closing the main window. There's probably a less convoluted and
safer way to do this, but I'm still not familiar enough with Makie's API
to implement it more elegantly.

Also, there appears to be some weird GLFW bug on exiting Julia if this
screen isn't destroyed, but I haven't investigated this enough to be
sure; it's cleaner to free it up anyway since we don't need it after
saving.