v0.8.0
GLMakie has been pinned to v0.8.6 as v0.8.7 includes a rewrite of the
inspector which causes our script to crash on mouseover of the fft image
so v0.8.6 will be our stable version for now.

I'm not sure if it's an issue with our script (since it's misusing a few
things) or if the rewrite introduced a new bug.

The crash is related to a missing interpolated_getindex method, which is
strange since our image isn't being interpolated (the `interpolate`
property is set to false) so it really should be calling
pixelated_getindex. But since I'm overriding the tooltip methods for the
inspector it's possible the issue is with the script. Either way,
something breaks with the newest version, so until I can figure out what
the problem is I'm going to leave the package pinned.

Of course this now necessitates instantiation and activation as anyone
with newer Makie versions will experience the crash unless they use this
project's environment.