Use the Makie API for setting the inspector tooltips
Recent versions of Makie (> v0.8.5) display an unnecessary third `z`
variable with a constant value of zero on the histogram tooltip when
using the DataInspector. Something changed during inspector refactoring
so that now histograms (a type of BarPlot from the looks of it) will
call the 3 dimensional method of position2text instead, bypassing our
override.

It's better to use their provided interface for this anyway.

The only drawback is that since the histogram is dynamically set with
user interaction (created and destroyed on Live toggle for performance
reasons) then we have to make sure this callback function is passed
every time the histogram is updated in this manner. In any case this
should still be easier to maintain and the aforementioned could be
easily solved with a much needed refactor.

Also, lines! now avoids setting a tooltip on hover.