Remove the Save figure button, use the array to save the FFT image
This partially reverts 9c9464d.

There's no need to jump through hoops to save the FFT as a figure when
we can just dump the normalized array. The result is grayscale rather
than colormapped, but this is preferable for the purposes of saving as
it's not only neutral but also encodes the intensity values directly.

For the figure since we're returning it it's simple enough to save, so
there's no need for a button.