Project.toml: fix the version string within the comment
Since this isn't a package and is meant to be used as a script these
specific key-value pairs can't be used because Julia expects a specific
directory structure and if I recall the code must also be embedded
within a module with a matching name. This is inconvenient, but these
fields are useful for tracking purposes and especially if I ever decide
to convert this into a package.

They must be commented out, however, so whenever any changes are
automatically made to the TOML file (such as adding another dependency)
then those changes must be brought in and staged patch style in order to
retain the comments.

Anyway, the `v` here was incorrect.