Remove a few nothings, add a return
This doesn't change the program's behaviour so it can be thought of as a
cosmetic change, but it's for consistency. Basically, the style enforced
here is if a function's return value is not used anywhere then the
function should explicitly return nothing. If it's an anonymous function
and the return value is not used (which in the context of this script is
usually a lambda within a `do` block) then the last expression (which
evaluates to the return value) should be `nothing` except when there's a
tail call and that subroutine is a procedure and thus its return is
void.

I remember this being useful when playing around with Observables and
ObserverFunctions which store these callback functions in their
`listeners` and `f` fields respectively and can be accessed and called
directly; having them return something unexpectedly should be avoided.

As for the function definitions, explicitly asserting a return yields
information at a glance in the sense that it tells you its return value
is not used anywhere. It's also a good habit to get into as it could
help type inference at the compiler level if the return type is
stable so it might avoid dynamic runtime dispatch in some cases.

There isn't universal consistency in that there's a distinction between
name bound functions and anonymous ones (we could just use `return`
everywhere), but the distinction is nice when reading through and
eliding the `nothing` with procedural tail calls also tells you the
expected return type of _that_ function if the convention is followed.