Redesign length observation switching
This genericizes the system and eliminates some unnecesary calls.
There are surely more efficient approaches, but this is simple enough
and gets the job done. The bulk of the inefficiency is at the tail end
for very large images -- dynamically processing large arrays through the
histogram is prohibitively expensive and that's not going to be solved
by micro-optimizing the length watch switches.