Make the statistics string an Observable
So I'm not sure why I didn't do this before since it's definitely
supported; or maybe it is now with newer versions of Makie and it wasn't
before, but I don't remember and it's more likely that I just did this
the hard way. Accessing the legend entry that way is definitely awkward.