Revert "Adjust the histogram axis limits"
This reverts commit 4de8530ff6f181c996d2e99f381ce9337a6ea3c6.

Fixing the x-limits here is not desirable as we want them to auto-adjust
depending on the selection (e.g. when few of the near-boundary values
are binned).

As for the y-limits, fixing the lower bound to zero creates a visual
asymmetry since now the x-limits are free, so the axis is now flush in
one direction but not the other.